LayoutTests/platform/chromium-win/fast/forms/calendar-picker/week-open-picker-with-f4-key-expected.txt

Tests if pressing F4 opens the calendar picker.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Check that page popup doesn't exist at first.
PASS document.getElementById("mock-page-popup") is null
Press F4 key.
PASS document.getElementById("mock-page-popup") is non-null.
PASS successfullyParsed is true

TEST COMPLETE
 
